+++ admin/src/Controller/Admin_fields_relationsController.php
@@ -217,7 +217,7 @@ class Admin_fields_relationsController extends FormController
 	 */
 	public function batch($model = null)
 	{
-		Session::checkToken() or jexit(Text::_('JINVALID_TOKEN'));
+		Session::checkToken() or exit(Text::_('JINVALID_TOKEN'));
 
 		// Set the model
 		$model = $this->getModel('Admin_fields_relations', '', []);