+++ admin/src/Controller/Component_admin_viewsController.php
@@ -217,7 +217,7 @@ class Component_admin_viewsController extends FormController
 	 */
 	public function batch($model = null)
 	{
-		Session::checkToken() or jexit(Text::_('JINVALID_TOKEN'));
+		Session::checkToken() or exit(Text::_('JINVALID_TOKEN'));
 
 		// Set the model
 		$model = $this->getModel('Component_admin_views', '', []);