+++ admin/src/Controller/Component_custom_admin_menusController.php
@@ -217,7 +217,7 @@ class Component_custom_admin_menusController extends FormController
 	 */
 	public function batch($model = null)
 	{
-		Session::checkToken() or jexit(Text::_('JINVALID_TOKEN'));
+		Session::checkToken() or exit(Text::_('JINVALID_TOKEN'));
 
 		// Set the model
 		$model = $this->getModel('Component_custom_admin_menus', '', []);