+++ admin/src/Controller/Component_files_foldersController.php
@@ -217,7 +217,7 @@ class Component_files_foldersController extends FormController
 	 */
 	public function batch($model = null)
 	{
-		Session::checkToken() or jexit(Text::_('JINVALID_TOKEN'));
+		Session::checkToken() or exit(Text::_('JINVALID_TOKEN'));
 
 		// Set the model
 		$model = $this->getModel('Component_files_folders', '', []);