+++ admin/src/Controller/Joomla_plugin_groupController.php
@@ -210,7 +210,7 @@ class Joomla_plugin_groupController extends FormController
 	 */
 	public function batch($model = null)
 	{
-		Session::checkToken() or jexit(Text::_('JINVALID_TOKEN'));
+		Session::checkToken() or exit(Text::_('JINVALID_TOKEN'));
 
 		// Set the model
 		$model = $this->getModel('Joomla_plugin_group', '', []);