+++ admin/src/Controller/Language_translationController.php
@@ -217,7 +217,7 @@ class Language_translationController extends FormController
 	 */
 	public function batch($model = null)
 	{
-		Session::checkToken() or jexit(Text::_('JINVALID_TOKEN'));
+		Session::checkToken() or exit(Text::_('JINVALID_TOKEN'));
 
 		// Set the model
 		$model = $this->getModel('Language_translation', '', []);