+++ admin/src/Controller/Library_files_folders_urlsController.php
@@ -217,7 +217,7 @@ class Library_files_folders_urlsController extends FormController
 	 */
 	public function batch($model = null)
 	{
-		Session::checkToken() or jexit(Text::_('JINVALID_TOKEN'));
+		Session::checkToken() or exit(Text::_('JINVALID_TOKEN'));
 
 		// Set the model
 		$model = $this->getModel('Library_files_folders_urls', '', []);