+++ admin/tmpl/admin_fields_conditions/modal.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * @package    Joomla.Component.Builder
+ *
+ * @created    30th April, 2015
+ * @author     Llewellyn van der Merwe <https://dev.vdm.io>
+ * @git        Joomla Component Builder <https://git.vdm.dev/joomla/Component-Builder>
+ * @copyright  Copyright (C) 2015 Vast Development Method. All rights reserved.
+ * @license    GNU General Public License version 2 or later; see LICENSE.txt
+ */
+
+// No direct access to this file
+defined('_JEXEC') or die;
+
+/** @var \VDM\Component\Componentbuilder\Administrator\View\Admin_fields_conditions\HtmlView $this */
+?>
+<div class="subhead noshadow mb-3">
+    <?php echo $this->getDocument()->getToolbar('toolbar')->render(); ?>
+</div>
+<div class="container-popup">
+    <?php $this->setLayout('edit'); ?>
+    <?php echo $this->loadTemplate(); ?>
+</div>