+++ admin/tmpl/joomla_plugin_files_folders_urls/default.php
@@ -24,6 +24,9 @@ Html::_('bootstrap.tooltip');
 // No direct access to this file
 defined('_JEXEC') or die;
 
+$layout  = $this->isModal ? 'modal' : 'edit';
+$tmpl    = $this->input->get('tmpl');
+$tmpl    = $tmpl ? '&tmpl=' . $tmpl : '';
 ?>
 <script type="text/javascript">
 	// waiting spinner
@@ -41,7 +44,7 @@ defined('_JEXEC') or die;
 	});
 </script>
 <div id="componentbuilder_loader" style="display: none;">
-<form action="<?php echo Route::_('index.php?option=com_componentbuilder&layout=edit&id='. (int) $this->item->id . $this->referral); ?>" method="post" name="adminForm" id="adminForm" class="form-validate" enctype="multipart/form-data">
+<form action="<?php echo Route::_('index.php?option=com_componentbuilder&&layout=' . $layout . $tmpl . '&id='. (int) $this->item->id . $this->referral); ?>" method="post" name="adminForm" id="adminForm" class="form-validate" enctype="multipart/form-data">
 
 <?php echo LayoutHelper::render('joomla_plugin_files_folders_urls.basic_above', $this); ?>
 <div class="main-card">